Code style updates, rename expressapp/controllers/index.js to expressapp/controllers/parkController.js
