+++ parkhangs-webapp/src/router/Routes.js
@@ -1,12 +1,12 @@
-import React, {Component} from 'react'
+import React, { Component  } from 'react'
 import NavBar from 'components/NavBar'
-import {Switch, Route} from 'react-router-dom'
+import { Switch, Route } from 'react-router-dom'
 import MapView from 'views/MapView'
 import CalendarsView from 'views/CalendarsView'
 import NoMatch from 'views/NoMatch'
 import Modal from 'features/modal/Modal'
 import 'App.css'
-import {connect} from "react-redux";
+import { connect } from "react-redux";
 import UserInfo from "../components/UserInfo";
 
 class Routes extends Component {
